backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
azure-cosmos==4.8.0
azure-storage-blob==12.24.0
azure-ai-vision-imageanalysis
azure-mgmt-notificationhubs==8.0.0
azure-identity==1.19.0
requests==2.32.3
pydantic==2.8.2
pydantic-settings==2.6.1
python-dateutil==2.9.0
Pillow==11.0.0
aiofiles==24.1.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0

